# OkSchool-main/lib/api-zod/src/generated/types/amazonProduct.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OkSchool API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AmazonProduct {
  id: number;
  title: string;
  description: string;
  affiliateUrl: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
  createdAt: Date;
}