# OkSchool-main/lib/api-zod/src/generated/types/amazonProductPatch.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OkSchool API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AmazonProductPatch {
  title?: string;
  description?: string;
  affiliateUrl?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
}