# OkSchool-main/lib/api-zod/src/generated/types/downloadEventInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OkSchool API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadEventInput {
  /** pdf or word */
  fileType: string;
}