# OkSchool-main/lib/api-zod/src/generated/types/mockTestDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OkSchool API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';

export interface MockTestDetail {
  id: number;
  title: string;
  description: string;
  subject: string;
  section: string;
  /** Duration in minutes */
  duration: number;
  questionCount: number;
  /** Marks awarded for a correct answer */
  correctMarks: number;
  /** Marks deducted for an incorrect answer */
  incorrectMarks: number;
  /** Marks for an unattempted question */
  unattemptedMarks: number;
  createdAt: Date;
  questions: Question[];
}